+++ src/modules/bin/into_module.rs
@@ -3,8 +3,9 @@ use std::collections::BTreeMap;
 use crate::{
     Environment, Expression, Int, LmError,
     expression::FileSize,
-    modules::bin::{
-        check_exact_args_len, from_module::parse_command_output, time_module::parse_time,
+    modules::{
+        bin::{check_exact_args_len, from_module::parse_command_output, time_module::parse_time},
+        pprint::strip_ansi_escapes,
     },
 };
 use common_macros::hash_map;
@@ -351,28 +352,4 @@ fn highlight_str(args: &[Expression], env: &mut Environment) -> Result<Expressio
 pub fn strip_str(args: &[Expression], env: &mut Environment) -> Result<Expression, LmError> {
     super::check_exact_args_len("striped", args, 1)?;
     Ok(strip_ansi_escapes(args[0].eval_in_assign(env)?.to_string().as_str()).into())
-}
-use regex_lite::Regex;
-pub fn strip_ansi_escapes(text: &str) -> String {
-    // 更全面的正则表达式，匹配大多数常见的 ANSI 转义序列
-    let ansi_escape_pattern = Regex::new(r"(?:\\x1b[@-_]|[\x80-\x9F])[0-?]*[ -/]*[@-~]").unwrap();
-    ansi_escape_pattern.replace_all(text, "").into_owned()
-    // (?:\\x1b[@-_]|[\x80-\x9F]):
-
-    // (?: ... )：这是一个非捕获组，表示匹配其中的内容但不捕获它。
-    // \\x1b[@-_]：匹配 \x1b 后面跟着 @ 到 _ 的字符。\x1b 是 ASCII 中的 ESC 字符（即转义字符），表示 ANSI 转义序列的开始。
-    // |：逻辑或操作符，表示匹配左边或右边的内容。
-    // [\x80-\x9F]：匹配从 \x80 到 \x9F 的字符范围。这些字符也是 ANSI 转义序列的一部分。
-    // [0-?]*:
-
-    // [0-?]：匹配从 0 到 ? 的字符范围。? 是 ASCII 中的一个特殊字符。
-    // *：表示前面的字符范围可以出现零次或多次。
-    // [ -/]*:
-
-    // [ -/]：匹配从空格到 / 的字符范围。
-    // *：表示前面的字符范围可以出现零次或多次。
-    // [@-~]:
-
-    // [@-~]：匹配从 @ 到 ~ 的字符范围。
-    // 这个范围包括了常见的控制字符，如 A-Z, a-z, 0-9, 和一些符号。
 }